style: Move Product entity to Product.entity package and ProductDTO move to Product.dto package
