+++ src/main/java/gift/product/ProductRepository.java
@@ -1,5 +1,6 @@
 package gift.product;
 
+import gift.product.entity.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
