+++ src/main/java/gift/wishlist/WishlistRepository.java
@@ -1,7 +1,7 @@
 package gift.wishlist;
 
 import gift.member.Member;
-import gift.product.Product;
+import gift.product.entity.Product;
 import java.util.List;
 import java.util.Optional;
 import org.springframework.data.domain.Page;
