+++ src/test/java/gift/option/OptionEntityTest.java
@@ -15,7 +15,7 @@ import gift.option.OptionTestCase.OptionNameAllowedCharacterError;
 import gift.option.OptionTestCase.OptionNameLengthError;
 import gift.option.OptionTestCase.OptionQuantitySizeError;
 import gift.option.entity.Option;
-import gift.product.Product;
+import gift.product.entity.Product;
 import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
