+++ src/test/java/gift/option/OptionServiceTest.java
@@ -21,7 +21,7 @@ import gift.option.OptionTestCase.OptionNameAllowedCharacterError;
 import gift.option.OptionTestCase.OptionNameLengthError;
 import gift.option.OptionTestCase.OptionQuantitySizeError;
 import gift.option.entity.Option;
-import gift.product.Product;
+import gift.product.entity.Product;
 import gift.product.ProductRepository;
 import java.util.List;
 import java.util.Optional;
