+++ src/test/java/gift/product/ProductRepositoryTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 
 import gift.category.Category;
 import gift.category.CategoryRepository;
+import gift.product.entity.Product;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
